Extract BIDS Dataset Class
